Usertroubleshooting

+++ server/schemas/resolvers.js
@@ -1,16 +1,13 @@
-const { AuthenticationError } = require('apollo-server-express');
-const { User, Todo } = require('../models');
-const { signToken } = require('../utils/auth');
+const { AuthenticationError } = require("apollo-server-express");
+const { User, Todo } = require("../models");
+const { signToken } = require("../utils/auth");
 
 const resolvers = {
   Query: {
-    users: async () => {
-      return User.find().populate('todos');
-    },
     user: async (parent, { username }) => {
-      return User.findOne({ username }).populate('todos');
+      return User.findOne({ username }).populate("todos");
     },
-    todos: async (parent, { username }) => {
+    todos: async (parent, { username }, { user }) => {
       const params = username ? { username } : {};
       return Todo.find(params).sort({ createdAt: -1 });
     },
@@ -23,45 +20,50 @@ const resolvers = {
     addUser: async (parent, args) => {
       const user = await User.create(args);
       const token = signToken(user);
-      console.log(user, token)
+      console.log(user, token);
       return { token, user };
     },
     login: async (parent, { email, password }) => {
       const user = await User.findOne({ email });
-
       if (!user) {
-        throw new AuthenticationError('No user found with this email address');
+        throw new AuthenticationError("No user found with this email address");
       }
-
       const correctPw = await user.isCorrectPassword(password);
-
       if (!correctPw) {
-        throw new AuthenticationError('Incorrect credentials');
+        throw new AuthenticationError("Incorrect credentials");
       }
-
       const token = signToken(user);
-
       return { token, user };
     },
-    addTodo: async (parent, { todoText, todoAuthor }) => {
-      const todo = await Todo.create({ todoText, todoAuthor });
-
+    addTodo: async (parent, { todoText, todoAuthor }, { user }) => {
+      if (!user) throw new AuthenticationError("You must be logged in");
+      const todo = await Todo.create({ todoText, todoAuthor: user.username });
       await User.findOneAndUpdate(
-        { username: todoAuthor },
+        { username: user.username },
         { $addToSet: { todos: todo._id } }
       );
-
       return todo;
     },
-    markTodoAsCompleted: async (parent, { todoId }) => {
-      const updatedTodo = await Todo.findOneAndUpdate(
+    markTodoAsCompleted: async (parent, { todoId }, { user }) => {
+      if (!user) throw new AuthenticationError("You must be logged in");
+      const todo = await Todo.findOne({
+        _id: todoId,
+        todoAuthor: user.username,
+      });
+      if (!todo) throw new AuthenticationError("Not authorized");
+      return Todo.findOneAndUpdate(
         { _id: todoId },
         { completed: true },
         { new: true }
       );
-      return updatedTodo;
     },
-    removeTodo: async (parent, { todoId }) => {
+    removeTodo: async (parent, { todoId }, { user }) => {
+      if (!user) throw new AuthenticationError("You must be logged in");
+      const todo = await Todo.findOne({
+        _id: todoId,
+        todoAuthor: user.username,
+      });
+      if (!todo) throw new AuthenticationError("Not authorized");
       return Todo.findOneAndDelete({ _id: todoId });
     },
   },
